Explicity mark Java sources with appropriate copyright information. This should have been done long ago and it recently came to my attention that it is only made more necessary since we have accepted contributions from the community.
Unless otherwise indicated, these sources have always been subject to the
Apache License, Version 2.0 as was stated on the Google Code project page
since its inception.  According to the License terms, all contributions
are subject to the same License unless otherwise indicated.  Most of the
contributions did not make any indication otherwise, which implies tacit
agreement with the License, Section 5, "Submission of Contributions."

  http://code.google.com/p/raphaelgwt/
  http://www.apache.org/licenses/LICENSE-2.0

It is indeed a sad day when such things are brought to light, but until
now the terms have been in effect anyway so I'd rather be explicit about
it than sweep it under the carpet.  Hurt me with the truth and all that.