Change AnimationCallback from an abstract class to an interface.
This commit breaks any classes that extend AnimationCallback;
you will need to change your class declaration to implement it
instead of extending it.  Aside from this simple change in the
declaration, user code is unaffected.

Note that instances of anonymous AnimationCallbacks are totally
unaffected by this change, because the syntax is identical for
creating new instances of anonymous classes that implement an
interface verses those that extend a class.  Funny thing about
Java, right there.